requirements.txt
flask>=3.0
requests>=2.31
Pillow>=10.0
pywebview>=5.3
pyinstaller>=6.0